Use float for all distances

+++ src/plafosim/message.py
@@ -167,8 +167,8 @@ class PlatoonAdvertisement(Message):
             platoon_speed: float,
             platoon_lane: int,
             platoon_formation: list,
-            platoon_position_front: int,
-            platoon_position_back: int):
+            platoon_position_front: float,
+            platoon_position_back: float):
         """Initialize a platoon advertisement instance"""
 
         super().__init__(origin, destination)
@@ -205,11 +205,11 @@ class PlatoonAdvertisement(Message):
         return self._platoon_formation
 
     @property
-    def platoon_position_front(self) -> int:
+    def platoon_position_front(self) -> float:
         return self._platoon_position_front
 
     @property
-    def platoon_position_back(self) -> int:
+    def platoon_position_back(self) -> float:
         return self._platoon_position_back
 
     def __str__(self) -> str:
